Skip unsupported NICs in DiscoverSriovDevices
Before creating the InterfaceExt in utils.DiscoverSriovDevices,
check that the discovered NIC is a supported model as configured
in the `supported-nic-ids` config map.

With this logic, the plugins will not get unsupported NIC to configure.

It is possible to get all the NICs discovered by using an env parameter:
DEV_MODE with value TRUE in the pod of the sriov-network-config-daemon.

Signed-off-by: Fred Rolland <frolland@nvidia.com>